+++ mitzasql/autocomplete/context_detection.py
@@ -1,3 +1,7 @@
+# Copyright (c) 2021 Vlad Balmos <vladbalmos@yahoo.com>
+# Author: Vlad Balmos <vladbalmos@yahoo.com>
+# See LICENSE file
+
 def detect_select_next_possible_keywords(context):
     if context == 'column':
         return ['from']