+++ mitzasql/autocomplete/smart_suggestions.py
@@ -1,3 +1,7 @@
+# Copyright (c) 2021 Vlad Balmos <vladbalmos@yahoo.com>
+# Author: Vlad Balmos <vladbalmos@yahoo.com>
+# See LICENSE file
+
 import bisect
 from mitzasql.utils import dfs, walk_ast
 from mitzasql.db.model import *