+++ mitzasql/sql_parser/ast.py
@@ -1,3 +1,7 @@
+# Copyright (c) 2021 Vlad Balmos <vladbalmos@yahoo.com>
+# Author: Vlad Balmos <vladbalmos@yahoo.com>
+# See LICENSE file
+
 simple_expr_unary_operators = ['+', '-', '~', '!']
 
 operator_precedance = {