+++ mitzasql/sql_parser/parser.py
@@ -1,3 +1,7 @@
+# Copyright (c) 2021 Vlad Balmos <vladbalmos@yahoo.com>
+# Author: Vlad Balmos <vladbalmos@yahoo.com>
+# See LICENSE file
+
 from .lexer import Lexer
 from .state import State
 from . import parser_factory