+++ mitzasql/sql_parser/parsers/parser.py
@@ -1,3 +1,7 @@
+# Copyright (c) 2021 Vlad Balmos <vladbalmos@yahoo.com>
+# Author: Vlad Balmos <vladbalmos@yahoo.com>
+# See LICENSE file
+
 class Parser:
     def __init__(self, state):
         self.state = state