+++ mitzasql/ui/ui.py
@@ -1,22 +1,6 @@
-# Copyright (c) 2019 Vlad Balmos <vladbalmos@yahoo.com>
+# Copyright (c) 2021 Vlad Balmos <vladbalmos@yahoo.com>
 # Author: Vlad Balmos <vladbalmos@yahoo.com>
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy of
-# this software and associated documentation files (the "Software"), to deal in
-# the Software without restriction, including without limitation the rights to
-# use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
-# the Software, and to permit persons to whom the Software is furnished to do so,
-# subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
-# FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
-# COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
-# IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
-# CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+# See LICENSE file
 
 import os
 import sys
@@ -88,7 +72,7 @@ def simulate_input_thread(fd, macro_file):
         line_counter = 1
         while True:
             line = file.readline()
-            if line is '':
+            if line == '':
                 break
 
             logger.info(u'Executing macro line {0}'.format(str(line_counter)))
@@ -167,4 +151,9 @@ def start_loop(macro_file=None):
     # in order to avoid circular reference in widgets when
     # importing the ui to refresh the screen
     shared_main_loop.mutable['loop'] = main_loop
-    main_loop.run()
+    old_signal_keys = None
+    try:
+        old_signal_keys = main_loop.screen.tty_signal_keys(intr='undefined')
+        main_loop.run()
+    finally:
+        main_loop.screen.tty_signal_keys(*old_signal_keys)